fix(backend): minor bug fix

+++ src/main/objects/server/PostTools.py
@@ -48,12 +48,12 @@ class PostTools:
             ids = []
             for i in comment_ids:
                 ids.append(i['comment_id'])
-        
-        if not self.__db.connect():
-            return generateResult("Check your internet connection", "connection")
-        else:
-            self.__db.insert(
-            f"""DELETE FROM Reactions WHERE comment_id IN ({', '.join([str(i) for i in ids])});""")
+            if ids != []:
+                if not self.__db.connect():
+                    return generateResult("Check your internet connection", "connection")
+                else:
+                    self.__db.insert(
+                    f"""DELETE FROM Reactions WHERE comment_id IN ({', '.join([str(i) for i in ids])});""")
         
         if not self.__db.connect():
             return generateResult("Check your internet connection", "connection")
